+++ packages/tswebextension/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Removed
+- `quickFixesRules` property from MV3 configuration.
+- Static getter `isUserScriptsApiSupported`
+
+
 ### Added
 
 - Export `isUserScriptsApiSupported` function to check if the current browser